Add empty test harness
Following these recommendations [1], we're not going to include any unit
tests in the stack-graphs library's src directory.  Instead, all of our
tests will live in a single "integration" test named `test/it`.  This
means that our test cases will only be able to use the library's public
interface, but that should help us ensure that we design a good API!

[1] https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html