Add path-finding algorithm
We use our new arena-allocated List type to store the symbol and scope
stacks efficiently.  The path-finding data is stored in arenas, just
like the stack graph structure, but we put those arenas in a separate
type.  That will allow us to run multiple path-finding operations in
parallel which all consume the same (immutable, already constructed)
stack graph.  Maybe that will be useful?