Add cycle detection
This is a simple heuristic that limits the number of paths we process
with the same start and end nodes.  Given all of the paths that we
encounter with the same start and end nodes, we try to "keep" the ones
that are shorter.  Our current threshold is 4 paths.  That said, all of
the specifics about this heuristic are subject to change in the future.
The only API that we guarantee is that our CycleDetector is a black box
that says "yes" or "no" to individual paths.