Add arena-allocated deques
The Deque type is similar to the List type, but it allows you to add or
remove elements from either end.  (The List type is singly-linked, and
so can only be used in a LIFO pattern — adding and removing only from
the front.)  To do this while retaining the nice structural sharing of
List, we use a ReversibleList under the covers to store the content, and
keep track of whether the inner list is currently storing the deque's
content in forwards or backwards order.  Whenever you perform on
operation that requires the elements to be in the opposite order than it
currently is, we reverse the list.

Note that you can definitely use this deque in a way that will have very
poor performance — for instance, alternating calls to `push_front` and
`push_back`.  But for our expected use cases, this implementation should
work out fine.