Add arena-allocated reversible lists
These store a singly-linked list of elements of some type, just like the
existing List type.  ReversibleLists add to that the ability to easily
reverse the contents of the list.  In particular, there's a `reverse`
method that does the work for you.  But more importantly, it caches the
result of reversing the list, so that it takes no additional work to
re-reverse it back to its original contents.