Add comparison operations to all arena lists
These are "deep" comparisons, which dereference into the arena and
compare the contents of the actual list elements.  We also provide
`equals_with` and `cmp_with` variants, where you pass in a function
that's used to compare the elements, which is helpful if your element
types can't implement `Eq` and `Ord`.  (That's true of anything that
contains a handle to other arena-allocated data, since the builtin
traits don't let you pass in references to the arena holding the actual
data.)