Add symbol and scope stack bindings
Partial paths can have variables in their preconditions and
postconditions.  When trying to append a partial path to a path, you
"match" the partial path's precondition against the path's corresponding
stack.  Any variables in the precondition can "bind" parts of the path's
stack.  You then "apply" those bindings to the partial path's
postcondition.  Any variable references in the postcondition are
substituted with the stack contents that were bound from the
precondition.  The result is the path's new stack after having performed
the concatenation.