Add database of partial paths
This database maintains internal in-memory indexes to support all of the
lookups that we might need to perform during the path-stitching
algorithm.  Most users will probably have some external storage layer
holding partial paths, in which case you will be responsible for loading
into the `Database` instance all of the partial paths that are valid
extensions of some current path.