Add phased path stitching implementation
This is an implementation of the path stitching algorithm that's divided
up into "phases".  At the start of each phase, we process whatever
(possibly incomplete) paths are currently in the queue.  As we extend
those paths with partial paths, we queue up the newly extended paths to
be processed in the _next_ phase.  This phasing approach means that the
Database instance doesn't need to be pre-loaded with _all_ of the
partial paths that we might ever need to process.  Instead, you're
notified of each path in the phase _before_ it will be processed, giving
you a chance to add its extensions to the database before allowing the
next phase to proceed.