Add node IDs for root and jump-to nodes
We can use None for a NodeID's file for these singleton nodes; since
handles are non-zero, this is nicely space-efficient.  We just need to
choose specific “local ID” values for each of the singleton nodes, and
then we can use the existing NodeID type for all nodes in a stack graph.