Add C bindings for creating symbols
This patch adds some initial C bindings for our stack graph library.
Right now, the only thing you can do is add and retrieve symbols.
However, this sets the pattern that we will follow with all other types:
direct access to the underlying arena, a stable C representation for
each type, and adding data to the graph in batches to minimize the
number of C calls that are needed.

We keep the auto-generated header file checked into git.  The CI job
verifies that the header file is up-to-date with any changes to the C
bindings' implementation.