Implement managed storage for symbols and files
Instead of storing a standard `String` in our symbol and file instances,
we now store string data in arena-like buffers, similar to the approach
described in [1].  This further reduces the amount of heap allocation
that's needed for our stack graph data, and eliminates the second copy
of each string that we were storing previously.

As a nice side effect, the `Symbol` and `File` types no longer contain
any fields that don't have a guaranteed `repr(C)` representation, which
means we can use them directly in the forthcoming C bindings.

[1] https://matklad.github.io/2020/03/22/fast-simple-rust-interner.html