Add C bindings for creating nodes and edges
You provide an array of `sg_node` instances and we add them into the
stack graph.  If any of the nodes are invalid, we skip those and return
a null handle for them.  Same for edges, although duplicate edges aren't
considered an error.

This patch also adds a `TestGraph` helper type, and an implementation of
`CreateStackGraph` that calls the A API.  This lets us add a copy of the
`can_create_graph` test that exercises the C API instead of the Rust
API.