Add C binding for finding complete paths
This function will typically only be useful in test cases, but is a good
building block on the way to a C API for the full phased path stitching
algorithm.  If you can call this function via the FFI, and validate the
results, then everything involving the StackGraphs and Paths types is
working as expected.