Ensure that partial path content can be reversed
This patch ensures that the content of all of the partial paths returned
by the path-finding algorithm is available in both directions.  We don't
want the consumer of the C API to have to invoke any additional C
methods to be able to iterate through the stack and edge list contents
in either direction.