Add C binding for phased path stitching algorithm
There's a little bit of trickery in the C API so that we don't have to
expose the PathStitcher type, but still provide access to the array of
paths found in the previous phase without needing to invoke a C
function.  But otherwise this lets you write a pretty straightforward
while loop to “drive” the path stitching algorithm from outside of Rust!
The test case shows an example of what this would look like, including a
fake “external storage layer” that we lazily load partial paths from.