Refer to exported scopes by ID, not handle
“Push scoped symbol” nodes need to refer to an exported scope node,
which will be added to the scope stack that's attached to the symbol
that is pushed onto the symbol stack.  Before, you would pass in the
handle of the exported scope node.  Now, you pass in its ID.  Both of
these are unique identifiers, and so there's no danger of an ambiguous
node reference.  This does mean that we need to look up the node's
handle as part of the path-finding algorithm, since we are still storing
node handles in our scope stack representations.

The reason for this change is that before this patch, you had to add
nodes to a stack graph in separate passes: first you would add your
exported scope nodes, and only then would you add your pushed scope
symbol nodes.  You had to use this ordering so that you'd have a valid
handle to pass in to the pushed scope symbol constructor.  By using node
IDs, there is no constraint on the order that you add nodes — it's
perfectly fine to add the pushed scoped symbol node first, as long you
as eventually add the exported scope node with the desired ID before
performing any lookups.

It's important to not enforce a particular order on node construction
since users of the C API have to use a batched API to add stack graph
content.  With this change, C consumers can use a single call to
`sg_stack_graph_add_nodes`, including _all_ of the nodes that they want
to create, instead of having to choreograph two separate calls to the
function.