Use “collapsed” strings in C API
Symbols and files both consist only of a single string.  Before, the
C API functions for bulk-loading symbols and files expected you to pass
in an array of arrays of string content.  This patch updates the
functions to have you pass in a single `const char*`, containing the
content of everything you want to add _concatenated together_.
Consumers will already likely be building up the string content
dynamically, and so it's not harder to build that up into a single
string buffer instead of into an array of string pointers.  And the
result should both require fewer allocations in all languages, and
should play more nicely with Cgo in particular, which prohibits Go
pointers within Go pointers for any data that you pass into a C call.