Get or create nodes in C API
The C API's function for adding nodes to a stack graph is now a “get or
create” function.  If there's already a node in the stack graph with the
same ID, we now return the handle of the existing node, instead of a
null handle.  We've renamed the function to make it explicitly clear
that there's a change of behavior.