Return partial path handles in C API
The C API function for adding partial paths to a database now returns
the handles of those partial paths, and there's a new accessor function
that gives access to the underlying arena.  This will be helpful in test
cases that add partial path content manually, so that you can verify
that they were added as expected.