Fix singleton node IDs in C representation
Our Rust code used to return the ID of the singleton root and jump to
scope nodes by constructing them on the fly.  This meant that we weren't
ensuring that we were filling in those nodes' in-memory ID with the
correct values.  Code using the C API that tried to access the IDs would
therefore get incorrect values.

We now fill in the IDs correctly, and have updated the Rust code to
always return the ID as it appears in memory.