Take control of C representation of None values
Our C API already includes comments describing how null values are
stored for several types.  This patch adds test cases to verify that
that's actually how we're storing them in memory.  In several cases
(mostly involving attached scope lists), we weren't, because the Rust
compiler was choosing a different field deep down inside the struct to
sneak the None value into.  (In particular, anything with a `Deque` was
using the value 2 for the deque's `direction` field, instead of the
value 0 for its `list` field.)

Fixing the test cases required a new `controlled-option` crate, which
uses a new `Niche` trait to take control over which "niche" value is
used to represent a None value.  This lets us explicitly say, for
instance, that null deques should use the 0 value of its `list` field.

After this patch, the rule of thumb is that any type that has a
`#[repr(C)]` annotation must use `ControlledOption` instead of `Option`
for any of its optional fields and nested subfields.