Ensure edge lists are available in both directions
The forward path stitcher returns a list of paths that have been
encountered in the most recent phase.  In the C wrapper API, we need to
ensure that the content of each paths' edge list is available in either
direction, so that the C caller can iterate through them in either
direction.