Add explicit symbol stack variables
We had several comments that talked about how we used an “implicit”
symbol stack variable in our partial symbol stacks.  As we move towards
a path-stitching algorithm that operates entirely on partial stacks,
it's more important to have consistency between the implementations of
the two stacks.  So this patch adds in symbol stack variables as a
first-class type in the data model.