Apply offsets to symbol and scope stack variables
You can now apply an offset to the various partial stack data
structures, which causes the “value” or “index” of all of the variables
inside of them to be incremented by a certain amount.  We'll use then
when concatenating partial paths as a way of ensuring that the left- and
right-hand sides have non-overlapping sets of variables.