Ensure no overlap between two partial paths
When using the concatenation operation from the previous patch, you will
typically want to make sure that the left- and right-hand sides have no
overlapping variable references.  (It's fine if they do, but that will
require the variables that appear on both sides to bind to the same
values for the concatenation to succeed.)  This patch adds a function
that edits a partial path to ensure that it has no variable references
that also appear in some other partial path.