Apply partial bindings to partial stacks
This mimics the existing functions for applying _concrete_ bindings to a
partial stack.  In that case, the result is a _concrete_ stack, since in
a set of concrete bindings every variable maps to a concrete stack.  If
we ensure that every preexisting variable has a binding, then the result
cannot possibly have any more variables in it.

In the partial case, that is no longer true.  A variable can map to any
arbitrary _partial_ stack, which means there might still be variables in
the result.  The result of applying partial bindings is therefore a
_partial_ stack.  As part of that, we no longer have to (or should)
treat it as an error if a preexisting variable doesn't have a binding —
in that case, the variable reference carries through to the result.