Concatenate partial paths
This new function concatenates two partial paths together, producing a
new partial path.  The result has all of the “starting state” (start
node, preconditions) of the left-hand side, and the “ending state” (end
node, postconditions) of the right-hand side.  We have to unify the
“middle parts” — the postcondition of the left-hand side and the
precondition of the right-hand side.  The result of that unification (if
it succeeds) is a set of variable bindings, which we then substitute
into the result's pre- and postconditions.