Track lengths in all stacks and lists
We were tracking an explicit length in some of the stack and list types,
but not others.  Let's be consistent.

This patch also updates our cycle detection logic for partial paths to
be consistent with the logic for concrete paths, since we have partial
symbol stack lengths available to us.