Index partial paths by end node too
We have a nice set of indices that we use in our forward path stitching
algorithms, which look for extensions that we can add to the right side
of a partial path.  That means that we have to index each partial path
by its start node and precondition, since that's the side of the
extension that “joins up” with the existing candidate path.

In the reverse direction, we have to do the opposite — it's the end node
and postcondition of each extension that we care about.  That means we
have to add two more indices to our Database type.