Add InternedStrings
This lets you add arbitrary string content to a stack graph, which will
be interned and deduplicated just like we're already doing for symbols
and filenames.  We're going to use this soon for some supplemental
information about stack graph nodes that are associated with a range of
content in a source file.