Allocate an unused entry 0 in supplemental arenas
We had already made this change to `Arena`; this patch makes a similar
change to `SupplementalArena`.

Before, the first element added to an arena was stored in the first slot
in the underlying `Vec` — meaning that it had index 0 within the arena.
However, we're using `NonZeroU32` as our handle type, so that we can use
the zero value as `None` when working with optional handles.  That meant
that we would have to add 1 to the element's index to get its handle.

This works fine in Rust, where we can hide the internal details of the
arena, and ensure that the increment/decrement always happens in the
public API methods.  In our C bindings, however, it will be nicer if we
can use an element's index directly as its handle.  For this to work, we
must reserve element 0 in each arena for the value that will never be
dereferenced.  Because most of our types don't have a default value that
makes sense, we use `MaybeUninit` to allocate an uninitialized element
0.  This is safe since it's not possible to dereference from Rust.