Determine which nodes in a file are “local”
A node is local if it cannot possibly participate in a name binding that
leaves its file.  In a stack graph, that means that there is no sequence
of partial paths that connects the node to the root node.

We have two methods for “setting” the set of local nodes.  The first is
meant to be called at index time, and actually analyzes the partial
paths in the database to determine which nodes are local.  The second is
meant to be called at query time, where you've stored the locality
information in your storage layer, and just need to load that back into
our internal view of the stack graph.