Bound the amount of work performed in each phase
In our phased path stitching algorithms, you can now bound the total
amount of work that can be performed in each phase.

The mechanics of using the stitching classes are exactly the same —
check which paths were newly discovered in the previous phase, and load
any possible extensions of them into the database before starting the
next phase.

The limit is configurable, and defaults to the previous behavior of no
additional bound.  (There is always _some_ limit on the amount of work
performed, since we only process extensions of paths found in previous
phases.)