Don't add divergent partial paths to database
A divergent partial path is one that starts at the root node and has an
empty symbol stack precondition.  That empty precondition means that it
can be concatenated to _any_ path that currently ends at the root node —
including the result of that concatenation!  That gives us a divergence,
since we can continually prepend the path's postcondition to the current
symbol stack, forever.

This patch adds a check that ensures that we never add this kind of
partial path to a `Database` instance.  (Without this check, we would
end up allocating an unusably large amount of memory, since we're
currently representing empty lists with a `u32::max` arena handle!)