Replace `should_extend` predicate with preconditions without stack variables in `ForwardPartialPathStitcher`
Forward partial path stitching allowed a `should_extend` predicate to control which paths were
considered for extension. This was used to cut off paths where the precondition was strengthened
during extension. However, it is easy for a user to forget to set it corectly, possibly resulting
in non-termination.

This patch removed that predicate, and instead ensures that all initial paths have no stack variables
in the precondition. This has the same effect as what the predicate was used for, and cannot be forgotten
by the user.