Change CancellationFlag
Remove the explicit `AtomicUsize` exposure from `CancellationFlag`, and
replace it with a `check` method similar to our other crates. A separate
`TreeSitterCancellationFlag` struct is created to handle bridging between
the Rust trait and the usize that Tree-sitter expects. This makes it
easier to make cancellation flags more flexible, e.g. by allowing them
to be chained or nested, without having to figure out how to map that
onto a single usize.