Add support for resolving ::Foo scoped constants
A program's lexical scope has an edge to the ROOT NODE,
and an edge is added from the ROOT NODE to a newly added
pop symbol node "<::>" representing the "root scope".

The root scope node has an edge to the program's defs.

I considered an edge from the root scope to the program's lexical scope,
but that causes the resulting complete path to cycle,
and it also allows the root scope node access it doesn't strictly
require (at least so far), whereas the root scope node only
for now needs access to the program's definitions. This produces
a shorter path.