Reorganize program to global scope connectivity
The scoping that is visible in Ruby sources, as determined by the
nesting of classes and modules, is independent of the source file
organization. Therefore, we expose all top-level definitions in
the global scope directly. A program's lexical scope inherits the
global scope, thus having access to all available definitions in
the repository.

Additionally, the global scope is accessible via the <::> marker,
which is used to ensure a name of the form ::NAME is resolved from the
root and not from an intermediate definition that happens to be in
scope.

Later we may consider introducing a notion of a project to isolate
groups of files from each other. This would be modeled as definitions
and references between the global and the program scope.