Factor out identifiers into seperate rules
Identifiers (`constant` and `scope_resolution`) can appear in both
reference and definition positions. To distinguish between these cases,
the rules would need to include the context, i.e., the surrounding term.
This works poorly with the recursive nature of `scope_resolution`, which
can only be expressed generally with top-level rules. Including context
would require explicit unfolding of the recursion (up to a fixed depth).

Instead of the unfolding, we introduce top-level rules which create both
reference and definition nodes. The terms using these names can chose to
use the definition or the reference, or both, if they wish.