Support root variable definitions
Definitions in the root scope (with names starting with `::`) can appear
inside nested classes and modules. To support this, the `.defs` property
has been split into `.lexical_defs` and `.root_defs`. The former behaves
the same as `.defs` did, containing the definitions that are part of the
enclosing construct. The latter contains definitions that instead escape
that context and bubble up to the root.