Implement local variables and lexical scoping
Local variables are only visible in the block they are defined, and only
after they've been assigned. Statements are chained, by conencting their
`.lexical_scope` to the `.lexical_scope` and the `.local_defs` of the
preceding statement.

Since the chaining behavior of statements is the same in a program and in
a `body_statement`, this is factored out into shared stanzas.