Implement refinement
Refinements are part of modules that provide additional instance methods
on classes that can be brought into scope by `using` the module. Because
refinements are not globally available, method resolution must carry
scope of the call to allow finding refinements at the call site.

A detailed example of refinement with the corresponding stack graph is
included in `doc/refinement.png`.