Redesign `Appendable`
The `Appendable` trait was confusing two concerns: (a) abstracting over
things that can be added to partial paths, and (b) allowing the use of
handles by tracking where handles can be looked up, represented by the
`Ctx` type parameter.

The new design separates these concerns. The `Appendable` only abstracts
over things that can be added to partial paths (currently edges and other
partial paths). A new `Index` trait is used to map handles to appendables,
and passed any methods that might need to do that.

En passent, a bug in the cycle detector was fixed that failed to rename
variables, resuling in occasional panics.