Reduce path finding memory usage
Cycle detection was maintaining a list where the elements were either
`PartialPath`s or `Handle`s. The former are much larger than the latter,
resulting in a larger enum size, even though most values in the list
would be the small ones.

To solve this, an additional arena is introduced that holds the partial
paths values themselves. The list elements are now handles into that
arena.