Remove lifetime bounds from `SpanCalculator`
`SpanCalculator` no longer holds a reference to the source file that
you're calculating positions within.  Instead, you must pass the source
text each time you calculate a position.  This allows us to remove the
lifetime parameter from the type, which is needed to (easily) add Lua
bindings for the type.